+++ src/crawl4md/convert/preprocessing/__init__.py
@@ -1,2 +1,1 @@
-from .markdown import MarkdownPreprocessing
-
+from .markdown import MarkdownPreprocessing as MarkdownPreprocessing